:truck: feat: migrate bleedingEdge config; remove deprecated classes
