--- bleedingEdge.neon
@@ -1,45 +1,0 @@
-parameters:
-    bootstrapFiles:
-        - silverstripe-autoloader.php
-services:
-    -
-        class: Cambis\Silverstan\Analyser\ResultCache\ResultCacheMetaExtension\ConfigCacheMetaExtension
-        tags: [phpstan.resultCacheMetaExtension]
-    -
-        class: Cambis\Silverstan\Autoloader\Autoloader
-    -
-        class: Cambis\Silverstan\ConfigurationResolver\Contract\MiddlewareRegistryProviderInterface
-        factory: Cambis\Silverstan\ConfigurationResolver\MiddlewareRegistryProvider\LazyMiddlewareRegistryProvider
-    configCollectionFactory:
-        class: Cambis\Silverstan\ConfigurationResolver\ConfigCollectionFactory\ExperimentalLazyConfigCollectionFactory
-    -
-        class: Cambis\Silverstan\ConfigurationResolver\Middleware\PrivateStaticMiddleware
-        tags: [silverstan.configurationResolver.middleware]
-    -
-        class: Cambis\Silverstan\ConfigurationResolver\Middleware\ExtensionMiddleware
-        tags: [silverstan.configurationResolver.middleware]
-    -
-        class: Cambis\Silverstan\ConfigurationResolver\Middleware\InheritanceMiddleware
-        tags: [silverstan.configurationResolver.middleware]
-    -
-        class: Cambis\Silverstan\ClassManifest\ClassManifest
-        arguments:
-            excludedClasses:
-                - 'SilverStripe\Core\Manifest\ClassManifestErrorHandler'
-            includeTestOnly: %silverstan.includeTestOnly%
-            parser: @currentPhpVersionPhpParser
-    -
-        class: Cambis\Silverstan\FileCleaner\FileCleaner
-    -
-        class: Cambis\Silverstan\FileFinder\FileFinder
-        arguments:
-            includeTestOnly: %silverstan.includeTestOnly%
-    -
-        class: Cambis\Silverstan\NodeVisitor\TestOnlyFinderVisitor
-    -
-        class: Composer\ClassMapGenerator\ClassMapGenerator
-        arguments:
-            extensions: ['php']
-    -
-        class: Cambis\Silverstan\TypeResolver\TypeResolver\ExtensionOwnerMetaPropertyTypeResolver
-        tags: [silverstan.typeResolver.propertyTypeResolver]
